requirements.txt
alembic==1.13.2
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
filelock==3.15.4
Flask==3.0.3
Flask-Cors==5.0.0
Flask-GoogleStorage==0.2.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.7.1
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet
gunicorn==22.0.0
idna==3.7
intelhex==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe
packaging==24.1
platformdirs==4.2.2
proto-plus==1.24.0
protobuf>=3.19.5,<5.0.0
psycopg2==2.9.10
psycopg2-binary --only-binary :all:
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
SQLAlchemy==2.0.31
tenacity==8.5.0
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
Werkzeug==3.0.3
google-cloud-firestore==2.16.0
firebase-admin==6.5.0


# psycopg2-binary==2.9.9